wifi: mac80211: Allow DFS/CSA on a link if scan is ongoing on another link
Currently, if a scan is ongoing, -EBUSY is returned when DFS or a channel
switch is initiated. However, for Multi-Link Operation (MLO), multiple
hardware can be grouped under a single wiphy. In such cases, if a scan is
ongoing on one underlying hardware and DFS or a channel switch is requested
on a different underlying hardware, these operations can be allowed
simultaneously.

Add logic to check the underlying hardware used for the requested scan. If
the hardware on which DFS or a channel switch is requested is not being
used, the operations are allowed; otherwise, -EBUSY is returned.

For non-MLO scenarios, there is no change.

Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>
Signed-off-by: Raj Kumar Bhagat <quic_rajkbhag@quicinc.com>
Signed-off-by: Jeff Johnson <jeff.johnson@oss.qualcomm.com>